+++ OpenSim/Moco/MocoProblem.h
@@ -156,6 +156,15 @@ public:
     void setControlInfo(const std::string& name, const MocoBounds&,
             const MocoInitialBounds& = {}, const MocoFinalBounds& = {});
 
+    /// Set information about a single Input control variable in this phase.
+    /// Similar to setControlInfo(). The name for an Input control follows the 
+    /// format <InputController_path>/<Input_label> (e.g.,
+    /// "/my_input_controller/input_control_0"). Input labels for an 
+    /// InputController can be obtained using 
+    /// InputController::getInputControlLabels().
+    void setInputControlInfo(const std::string& name, const MocoBounds&,
+            const MocoInitialBounds& = {}, const MocoFinalBounds& = {});
+
     /// Set the bounds on generalized speed state variables
     /// for which explicit bounds are not set.
     void setDefaultSpeedBounds(const MocoBounds& bounds) {
@@ -166,11 +175,22 @@ public:
     /// activations, etc. Infos provided via setControlInfoPattern() take
     /// precedence over the default values from the model. Infos provided via
     /// setControlInfo() take precedence over infos provided with
-    /// setControlInfoPattern().  If a state variable name matches multiple
+    /// setControlInfoPattern(). If a control variable name matches multiple
     /// patterns, the info provided with the last pattern is used for that
     /// control variable.
     void setControlInfoPattern(const std::string& pattern, const MocoBounds&,
             const MocoInitialBounds& = {}, const MocoFinalBounds& = {});
+
+    /// Set information for Input control variables whose names match the 
+    /// provided regular expression. Infos provided via 
+    /// setInputControlInfo() take precedence over infos provided with 
+    /// setInputControlInfoPatter(). If an Input control variable name matches
+    /// multiple patterns, the info provided with the last pattern is used for
+    /// that Input control variable.
+    void setInputControlInfoPattern(const std::string& pattern, 
+            const MocoBounds&, const MocoInitialBounds& = {}, 
+            const MocoFinalBounds& = {});
+    
     /// For muscles without explicit activation bounds, set the bounds for
     /// muscle activation (if activation dynamics are enabled) from the bounds
     /// for muscle control (excitation), using min/max control if explicit
@@ -376,6 +396,11 @@ protected: // Protected so that doxygen shows the properties.
     OpenSim_DECLARE_LIST_PROPERTY(control_infos_pattern, MocoVariableInfo,
             "Set control variable bounds for all controls matching a regular "
             "expression.");
+    OpenSim_DECLARE_LIST_PROPERTY(input_control_infos, MocoVariableInfo, 
+            "The Input control variables' bounds.");
+    OpenSim_DECLARE_LIST_PROPERTY(input_control_infos_pattern, MocoVariableInfo,
+            "Set control variable bounds for all Input controls matching a "
+            "regular expression.");
     OpenSim_DECLARE_LIST_PROPERTY(parameters, MocoParameter,
             "Parameter variables (model properties) to optimize.");
     OpenSim_DECLARE_LIST_PROPERTY(goals, MocoGoal,
@@ -462,6 +487,13 @@ public:
     /// Set bounds for a control variable using a regular expression.
     void setControlInfoPattern(const std::string& pattern, const MocoBounds&,
             const MocoInitialBounds& = {}, const MocoFinalBounds& = {});
+    /// Set bounds for an Input control variable for phase 0.
+    void setInputControlInfo(const std::string& name, const MocoBounds&,
+            const MocoInitialBounds& = {}, const MocoFinalBounds& = {});
+    /// Set bounds for an Input control variable using a regular expression.
+    void setInputControlInfoPattern(const std::string& pattern, 
+            const MocoBounds&, const MocoInitialBounds& = {}, 
+            const MocoFinalBounds& = {});
     /// Set bounds for the kinematic constraints in phase 0.
     void setKinematicConstraintBounds(const MocoBounds& bounds);
     /// Set bounds for the Lagrange multipliers in phase 0.